feat(a11y): add ARIA attribute generation and modularize accessibility module
This implements issue #165 - ARIA attributes and semantic markup support.

Changes:
- Add ARIA attribute generation utilities (aria.rs)
  - AriaAttribute enum for all standard ARIA attributes
  - AriaBuilder for constructing ARIA attributes from AccessibleNode
  - LiveRegion support for aria-live and aria-atomic
  - Helper functions: aria_attributes(), aria_pairs()

- Modularize accessibility module:
  - Split accessibility.rs into focused module files:
    - roles.rs: Role enum and implementation
    - state.rs: AccessibleState
    - node.rs: AccessibleNode
    - announcement.rs: Announcement and Priority
    - manager.rs: AccessibilityManager and SharedAccessibility
    - aria.rs: ARIA attribute generation
  - All public APIs preserved through mod.rs re-exports
  - Tests moved to separate tests.rs file

- Enhanced accessibility features:
  - aria-label, aria-labelledby, aria-describedby support
  - aria-live regions (polite/assertive/off)
  - aria-atomic for live regions
  - All ARIA state attributes (expanded, checked, selected, etc.)
  - Value range attributes (valuemin, valuemax, valuenow)
  - Position in set attributes (posinset, setsize)
  - Tabindex support based on focusability

This provides a foundation for semantic markup in TUI applications
and enables better screen reader integration.
